added hamburger menu, general fixes and style changes
hamburger menu for mobile devices in place. Some style changes and
structural redesign within menu due, but menu itself is done.
experimenting with social media icons. Being a bitch.
seperated sideNav html code to a clipboard file for later use and to
tidy up the html a bit.
Tried to do favicon again, but for some reason I'm incapable of that. :/
downloaded some sample images for directory gallery when it comes into
effect.