I don't even know anymore
The projects page toggleProjectNav function has stopped existing
according to chrome. I disagree, but it's not like I can exactly go have
a conversation with it now, is it?

scrollRight and scrollLeft continue to refuse to follow the rules as
presented bby the MDN, so again I'm kinda lost

Bright side is that tutorials page looks a little more like what I
envisioned, i.e. it has something on the damn page now. Just gotta work
out how to make the video expand width when I click on it in js, because
the css solution only works so long as you hold the click on it :(

other miscellanous shite goes here if there is any to commit